JAMES-1644 Add guice injection in JMAP module
git-svn-id: https://svn.apache.org/repos/asf/james/project/trunk@1719323 13f79535-47bb-0310-9956-ffa450edef68

+++ server/data/data-jmap/src/test/java/org/apache/james/jmap/memory/access/MemoryAccessTokenRepositoryTest.java
@@ -32,7 +32,7 @@ public class MemoryAccessTokenRepositoryTest {
 
     private static final AccessToken TOKEN = AccessToken.generate();
     private static final String USERNAME = "username";
-    private static final int TTL_IN_MS = 100;
+    private static final long TTL_IN_MS = 100;
 
     private MemoryAccessTokenRepository accessTokenRepository;
